# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DropFlow - Dropshipping Automation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Add a hunted product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/generate-description:
    post:
      operationId: generateProductDescription
      tags: [products]
      summary: AI-generate a product description
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  # ---------- SUPPLIERS ----------
...
  /reports/pl:
    get:
      operationId: getPlReport
      tags: [reports]
      summary: Generate a Profit & Loss report for a date range
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
          description: ISO date string — start of range (inclusive)
        - name: to
          in: query
          required: true
          schema:
            type: string
          description: ISO date string — end of range (inclusive)
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [product, supplier, status]
          description: Dimension to group rows by (default product)
      responses:
        "200":
          description: P&L report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlReport"

  # ---------- PRICE WATCH ----------
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: ["string", "null"]
        niche:
          type: ["string", "null"]
        costPrice:
          type: ["number", "null"]
        sellPrice:
          type: ["number", "null"]
        margin:
          type: ["number", "null"]
        status:
          type: string
          enum: [hunting, researching, listed, archived]
        description:
          type: ["string", "null"]
        aiDescription:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        supplierId:
          type: ["integer", "null"]
        stockQuantity:
          type: ["integer", "null"]
        stockThreshold:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ProductInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        niche:
          type: string
        costPrice:
          type: number
        sellPrice:
          type: number
        status:
          type: string
          enum: [hunting, researching, listed, archived]
        description:
          type: string
        imageUrl:
          type: string
        sourceUrl:
          type: string
        notes:
          type: string
        supplierId:
          type: integer
        stockQuantity:
          type: integer
        stockThreshold:
          type: integer
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        niche:
          type: string
        costPrice:
          type: number
        sellPrice:
          type: number
        status:
          type: string
          enum: [hunting, researching, listed, archived]
        description:
          type: string
        aiDescription:
          type: string
        imageUrl:
          type: string
        sourceUrl:
          type: string
        notes:
          type: string
        supplierId:
          type: integer
        stockQuantity:
          type: integer
        stockThreshold:
          type: integer

    # ---- Supplier ----
...
    PlReportRow:
      type: object
      required: [label, orderCount, revenue, cogs, grossProfit, margin]
      properties:
        label:
          type: string
        orderCount:
          type: integer
        revenue:
          type: number
        cogs:
          type: number
        grossProfit:
          type: number
        margin:
          type: number
          description: Gross margin percentage
    PlReport:
      type: object
      required:
        [
          from,
          to,
          groupBy,
          rows,
          totalRevenue,
          totalCogs,
          totalGrossProfit,
          totalMargin,
          totalOrders,
        ]
      properties:
        from:
          type: string
        to:
          type: string
        groupBy:
          type: string
        rows:
          type: array
          items:
            $ref: "#/components/schemas/PlReportRow"
        totalRevenue:
          type: number
        totalCogs:
          type: number
        totalGrossProfit:
          type: number
        totalMargin:
          type: number
        totalOrders:
          type: integer

    # ---- Price Watch ----
...